+++ interface.go
@@ -186,7 +186,7 @@ type Connection interface {
 	ReceiveMessage() ([]byte, error)
 
 	// StartRateShaping signals the connection to starts sending data at a constant rate.
-	StartRateShaping(int)
+	StartRateShaping(float64)
 	// StartPacketShaping signals the connection to send always fixed-sized packets.
 	StartPacketShaping(int)
 	// PauseSending signals the connection to halts all transmission.
